Splits up GrowthRequirement into 2 classes.
Currently it has 2 responsibilities.
Also replaces soil lists with sets.
