+++ src/main/java/com/InfinityRaider/AgriCraft/gui/GuiJournal.java
@@ -1,6 +1,6 @@
 package com.InfinityRaider.AgriCraft.gui;
 
-import com.InfinityRaider.AgriCraft.farming.GrowthRequirement;
+import com.InfinityRaider.AgriCraft.farming.GrowthRequirements;
 import com.InfinityRaider.AgriCraft.farming.mutation.Mutation;
 import com.InfinityRaider.AgriCraft.farming.mutation.MutationHandler;
 import com.InfinityRaider.AgriCraft.items.ItemJournal;
@@ -10,12 +10,10 @@ import com.InfinityRaider.AgriCraft.utility.BlockWithMeta;
 import com.InfinityRaider.AgriCraft.utility.IOHelper;
 import com.InfinityRaider.AgriCraft.utility.RenderHelper;
 import com.InfinityRaider.AgriCraft.utility.SeedHelper;
-import net.minecraft.block.Block;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiScreen;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.init.Blocks;
 import net.minecraft.item.ItemBlock;
 import net.minecraft.item.ItemSeeds;
 import net.minecraft.item.ItemStack;
@@ -384,7 +382,7 @@ public class GuiJournal extends GuiScreen {
     private void drawSeedTitle(int index) {
         ItemStack seed = discoveredSeeds[index];
         String title = seed.getDisplayName();
-        BlockWithMeta soil = GrowthRequirement.getGrowthRequirement((ItemSeeds) seed.getItem(), seed.getItemDamage()).getSoil();
+        BlockWithMeta soil = GrowthRequirements.getGrowthRequirement((ItemSeeds) seed.getItem(), seed.getItemDamage()).getSoil();
         Minecraft.getMinecraft().getTextureManager().bindTexture(RenderHelper.getItemResource(seedIcon));
         if(soil!=null) {
             this.renderIconInGui(this.guiLeft + 26, this.guiTop + 11, RenderHelper.getBlockResource(soil.getBlock().getIcon(1, soil.getMeta())));
